Inline the createValidSierraTransformableJsonWith helper
