+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/platform/transformer/SierraTransformerFeatureTest.scala
@@ -1,13 +1,13 @@
 package uk.ac.wellcome.platform.transformer
 
+import io.circe.KeyEncoder
 import org.scalatest.concurrent.Eventually
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.messaging.test.fixtures.{Messaging, SNS, SQS}
-import uk.ac.wellcome.models.work.internal.{
-  IdentifierType,
-  SourceIdentifier,
-  UnidentifiedWork
-}
+import uk.ac.wellcome.models.transformable.SierraTransformable
+import uk.ac.wellcome.models.transformable.sierra.SierraItemNumber
+import uk.ac.wellcome.models.transformable.sierra.test.utils.SierraUtil
+import uk.ac.wellcome.models.work.internal.{IdentifierType, SourceIdentifier, UnidentifiedWork}
 import uk.ac.wellcome.storage.fixtures.S3
 import uk.ac.wellcome.test.utils.ExtendedPatience
 import uk.ac.wellcome.platform.transformer.utils.TransformableMessageUtils
@@ -23,7 +23,10 @@ class SierraTransformerFeatureTest
     with fixtures.Server
     with Eventually
     with ExtendedPatience
+    with SierraUtil
     with TransformableMessageUtils {
+
+  implicit val keyEncoder: KeyEncoder[SierraItemNumber] = SierraTransformable.keyEncoder
 
   it("transforms sierra records and publishes the result to the given topic") {
     val id = createSierraBibNumber
@@ -33,12 +36,25 @@ class SierraTransformerFeatureTest
       withLocalSqsQueue { queue =>
         withLocalS3Bucket { storageBucket =>
           withLocalS3Bucket { messagingBucket =>
+            val data =
+              s"""
+                 |{
+                 | "id": "$id",
+                 | "title": "$title",
+                 | "varFields": []
+                 |}
+                  """.stripMargin
+
+            val sierraTransformable = SierraTransformable(
+              bibRecord = createSierraBibRecordWith(
+                id = id,
+                data = data
+              )
+            )
+
             val sierraHybridRecordMessage =
               hybridRecordNotificationMessage(
-                message = createValidSierraTransformableJsonWith(
-                  id = id,
-                  title = title
-                ),
+                message = toJson(sierraTransformable).get,
                 sourceName = "sierra",
                 s3Client = s3Client,
                 bucket = storageBucket
