+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/platform/transformer/utils/TransformableMessageUtils.scala
@@ -1,41 +1,15 @@
 package uk.ac.wellcome.platform.transformer.utils
 
 import com.amazonaws.services.s3.AmazonS3
-import io.circe.KeyEncoder
 import uk.ac.wellcome.messaging.test.fixtures.SQS
-import uk.ac.wellcome.models.transformable.sierra.{SierraBibNumber, SierraItemNumber}
-import uk.ac.wellcome.models.transformable.sierra.test.utils.SierraUtil
-import uk.ac.wellcome.models.transformable.{MiroTransformable, SierraTransformable}
+import uk.ac.wellcome.models.transformable.MiroTransformable
+import uk.ac.wellcome.models.work.test.util.IdentifiersUtil
 import uk.ac.wellcome.storage.fixtures.S3.Bucket
 import uk.ac.wellcome.storage.vhs.{HybridRecord, SourceMetadata}
 import uk.ac.wellcome.utils.JsonUtil
 import uk.ac.wellcome.utils.JsonUtil._
 
-trait TransformableMessageUtils extends SierraUtil with SQS {
-  implicit val keyEncoder: KeyEncoder[SierraItemNumber] = SierraTransformable.keyEncoder
-
-  def createValidSierraTransformableJsonWith(id: SierraBibNumber =
-                                               createSierraBibNumber,
-                                             title: String): String = {
-    val data =
-      s"""
-         |{
-         | "id": "$id",
-         | "title": "$title",
-         | "varFields": []
-         |}
-      """.stripMargin
-
-    val sierraTransformable = SierraTransformable(
-      bibRecord = createSierraBibRecordWith(
-        id = id,
-        data = data
-      )
-    )
-
-    JsonUtil.toJson(sierraTransformable).get
-  }
-
+trait TransformableMessageUtils extends IdentifiersUtil with SQS {
   def createValidMiroTransformableJson(MiroID: String,
                                        MiroCollection: String,
                                        data: String): String = {
